backend/requirements.txt
Django==2.0.2
django-rest-auth==0.9.3
django-rest-framework==0.1.0
djangorestframework==3.7.7
pytz==2018.3
six==1.11.0
